@@ -1,12 +1,12 @@
 package filter
 
 import (
-	"path/filepath"
-	"runtime"
-	"testing"
 	"github.com/jarcoal/httpmock"
 	"github.com/layer5io/meshery/mesheryctl/pkg/utils"
 	"github.com/stretchr/testify/assert"
+	"path/filepath"
+	"runtime"
+	"testing"
 )
 
 func TestListCmd(t *testing.T) {
