@@ -3,15 +3,15 @@ package handlers
 import (
 	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
-	"strings"
 	"github.com/gofrs/uuid"
 	"github.com/gorilla/mux"
 	"github.com/layer5io/meshery/server/meshes"
 	"github.com/layer5io/meshery/server/models"
 	"github.com/layer5io/meshkit/errors"
 	"github.com/layer5io/meshkit/models/events"
+	"io"
+	"net/http"
+	"strings"
 
 	pCore "github.com/layer5io/meshery/server/models/pattern/core"
 	"github.com/layer5io/meshery/server/models/pattern/stages"
@@ -119,8 +119,8 @@ func (h *Handler) handlePatternPOST(
 			rw.WriteHeader(http.StatusBadRequest)
 			fmt.Fprintf(rw, "%s", err)
 			event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-			"error": ErrSavePattern(err),
-		}).WithDescription("Pattern save failed, cytoJSON could be malformed.").Build()
+				"error": ErrSavePattern(err),
+			}).WithDescription("Pattern save failed, cytoJSON could be malformed.").Build()
 
 			_ = provider.PersistEvent(event)
 			go h.config.EventBroadcaster.Publish(userID, event)
@@ -132,7 +132,7 @@ func (h *Handler) handlePatternPOST(
 			rw.WriteHeader(http.StatusInternalServerError)
 			fmt.Fprintf(rw, "%s", err)
 			event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-			"error": ErrSavePattern(err),
+				"error": ErrSavePattern(err),
 			}).WithDescription(ErrSavePattern(err).Error()).Build()
 
 			_ = provider.PersistEvent(event)
@@ -145,7 +145,7 @@ func (h *Handler) handlePatternPOST(
 			h.log.Error(ErrSavePattern(err))
 			http.Error(rw, ErrSavePattern(err).Error(), http.StatusBadRequest)
 			event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-			"error": ErrSavePattern(err),
+				"error": ErrSavePattern(err),
 			}).WithDescription("unable to get \"name\" from the pattern.").Build()
 
 			_ = provider.PersistEvent(event)
@@ -161,17 +161,19 @@ func (h *Handler) handlePatternPOST(
 				"path": "",
 				"type": "local",
 			},
+			CatalogData: parsedBody.PatternData.CatalogData,
 		}
 		if parsedBody.PatternData != nil {
 			mesheryPattern.ID = parsedBody.PatternData.ID
 		}
+
 		if parsedBody.Save {
 			resp, err := provider.SaveMesheryPattern(token, mesheryPattern)
 			if err != nil {
 				h.log.Error(ErrSavePattern(err))
 				http.Error(rw, ErrSavePattern(err).Error(), http.StatusInternalServerError)
 				event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-				"error": ErrSavePattern(err),
+					"error": ErrSavePattern(err),
 				}).WithDescription(ErrSavePattern(err).Error()).Build()
 
 				_ = provider.PersistEvent(event)
@@ -179,7 +181,6 @@ func (h *Handler) handlePatternPOST(
 				return
 			}
 
-			
 			h.formatPatternOutput(rw, resp, format, &res, eventBuilder)
 			event := eventBuilder.Build()
 			_ = provider.PersistEvent(event)
@@ -199,7 +200,7 @@ func (h *Handler) handlePatternPOST(
 		}
 
 		h.formatPatternOutput(rw, byt, format, &res, eventBuilder)
-	
+
 		return
 	}
 	// If Content is not empty then assume it's a local upload
@@ -220,7 +221,7 @@ func (h *Handler) handlePatternPOST(
 				h.log.Error(ErrSavePattern(err))
 				http.Error(rw, ErrSavePattern(err).Error(), http.StatusBadRequest)
 				event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-				"error": ErrSavePattern(err),
+					"error": ErrSavePattern(err),
 				}).WithDescription("unable to get \"name\" from the pattern.").Build()
 
 				_ = provider.PersistEvent(event)
@@ -247,9 +248,9 @@ func (h *Handler) handlePatternPOST(
 			if err != nil {
 				h.log.Error(ErrSavePattern(err))
 				http.Error(rw, ErrSavePattern(err).Error(), http.StatusInternalServerError)
-				
+
 				event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-				"error": ErrSavePattern(err),
+					"error": ErrSavePattern(err),
 				}).WithDescription(ErrSavePattern(err).Error()).Build()
 
 				_ = provider.PersistEvent(event)
@@ -288,7 +289,7 @@ func (h *Handler) handlePatternPOST(
 			h.log.Error(ErrImportPattern(err))
 			http.Error(rw, ErrImportPattern(err).Error(), http.StatusInternalServerError)
 			event := eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-			"error": ErrImportPattern(err),
+				"error": ErrImportPattern(err),
 			}).WithDescription(ErrImportPattern(err).Error()).Build()
 
 			_ = provider.PersistEvent(event)
@@ -730,9 +731,9 @@ func (h *Handler) formatPatternOutput(rw http.ResponseWriter, content []byte, fo
 			patternFile, err := pCore.NewPatternFile([]byte(content.PatternFile))
 			if err != nil {
 				http.Error(rw, ErrParsePattern(err).Error(), http.StatusBadRequest)
-				
+
 				eventBuilder.WithSeverity(events.Error).WithMetadata(map[string]interface{}{
-				"error": ErrParsePattern(err),
+					"error": ErrParsePattern(err),
 				}).WithDescription("Unable to parse pattern file, pattern could be malformed.").Build()
 				return
 			}
