@@ -12,11 +12,11 @@ func GetAddonsState(ctx context.Context, selectors []MeshType, provider models.P
 	addonlist := make([]*AddonList, 0)
 	objects := make([]meshsyncmodel.Object, 0)
 	cidMap := make(map[string]bool)
-	
+
 	for _, c := range cid {
 		cidMap[c] = true
 	}
-	
+
 	for _, selector := range selectors {
 		//subquery1 := r.DBHandler.Select("id").Where("kind = ? AND key = ? AND value = ?", meshsyncmodel.KindAnnotation, "meshery/component-type", "control-plane").Table("key_values")
 		//subquery2 := r.DBHandler.Select("id").Where("id IN (?) AND kind = ? AND key = ? AND value IN (?)", subquery1, meshsyncmodel.KindAnnotation, "meshery/maintainer", selectors).Table("key_values")
