@@ -20,7 +20,7 @@ type MesheryFilter struct {
 
 	Location       sql.Map    `json:"location"`
 	Visibility     string     `json:"visibility"`
-	CatalogData    sql.Map    `json:"catalogData"`
+	CatalogData    sql.Map    `json:"catalog_data"`
 	FilterResource string     `json:"filter_resource"`
 	UpdatedAt      *time.Time `json:"updated_at,omitempty"`
 	CreatedAt      *time.Time `json:"created_at,omitempty"`
@@ -37,7 +37,7 @@ type MesheryFilterPayload struct {
 
 	Location       sql.Map    `json:"location"`
 	Visibility     string     `json:"visibility"`
-	CatalogData    sql.Map    `json:"catalogData"`
+	CatalogData    sql.Map    `json:"catalog_data"`
 	FilterResource string     `json:"filter_resource"`
 	Config         string     `json:"config"`
 	UpdatedAt      *time.Time `json:"updated_at,omitempty"`
