@@ -20,7 +20,7 @@ type MesheryPattern struct {
 
 	Location    sql.Map `json:"location"`
 	Visibility  string  `json:"visibility"`
-	CatalogData sql.Map `json:"catalogData"`
+	CatalogData sql.Map `json:"catalog_data,omitempty"`
 
 	UpdatedAt *time.Time `json:"updated_at,omitempty"`
 	CreatedAt *time.Time `json:"created_at,omitempty"`
