@@ -94,6 +94,7 @@ const styles = (theme) => ({
     paddingLeft: 34,
     paddingRight: 34,
     backgroundColor: theme.palette.secondary.mainBackground,
+    boxShadow: `3px 0px 4px ${theme.palette.secondary.focused}`,
   },
   toolbarOnDrawerOpen: {
     minHeight: 58,
@@ -102,6 +103,7 @@ const styles = (theme) => ({
     paddingRight: 34,
     backgroundColor: theme.palette.secondary.mainBackground,
     [theme.breakpoints.between(620, 732)]: { minHeight: 68, paddingLeft: 20, paddingRight: 20 },
+    boxShadow: `3px 0px 4px ${theme.palette.secondary.focused}`,
   },
   itemActiveItem: { color: '#00B39F' },
   headerIcons: { fontSize: '1.5rem', height: '1.5rem', width: '1.5rem' },
@@ -146,7 +148,7 @@ const styles = (theme) => ({
     marginTop: '-0.7rem',
     borderRadius: '3px',
     padding: '1rem',
-    zIndex: 1201,
+    // zIndex: 1201,
     boxShadow: '20px #979797',
     transition: 'linear .2s',
     transitionProperty: 'height',
@@ -258,7 +260,7 @@ function K8sContextMenu({
   const { notify } = useNotification();
   const styleSlider = {
     position: 'absolute',
-    left: '-5rem',
+    left: '-7rem',
     zIndex: '-1',
     bottom: showFullContextMenu ? '-55%' : '-110%',
     transform: showFullContextMenu ? `translateY(${transformProperty}%)` : 'translateY(0)',
@@ -392,7 +394,7 @@ function K8sContextMenu({
               src="/static/img/kubernetes.svg"
               width="24px"
               height="24px"
-              style={{ zIndex: '2' }}
+              // style={{ zIndex: '2' }}
             />
             <div className={classes.cbadge}>{contexts?.total_count || 0}</div>
           </div>
@@ -594,7 +596,7 @@ class Header extends React.PureComponent {
           <AppBar
             color="primary"
             position="sticky"
-            elevation={2}
+            // elevation={1}
             className={onDrawerCollapse ? classes.appBarOnDrawerClosed : classes.appBarOnDrawerOpen}
           >
             <Toolbar
