@@ -1,5 +1,6 @@
 import {
   Avatar,
+  Box,
   Button,
   Dialog,
   DialogActions,
@@ -66,6 +67,10 @@ import CustomColumnVisibilityControl from '../utils/custom-column';
 import ResponsiveDataTable from '../utils/data-table';
 import useStyles from '../assets/styles/general/tool.styles';
 import { Edit as EditIcon } from '@material-ui/icons';
+import { updateVisibleColumns } from '../utils/responsive-column';
+import { useWindowDimensions } from '../utils/dimension';
+import InfoModal from './Modals/Information/InfoModal';
+import InfoOutlinedIcon from '@material-ui/icons/InfoOutlined';
 
 const styles = (theme) => ({
   grid: {
@@ -265,6 +270,7 @@ function YAMLEditor({ pattern, onClose, onSubmit }) {
                 id: pattern.id,
                 name: pattern.name,
                 type: FILE_OPS.UPDATE,
+                catalog_data: pattern.catalog_data,
               })
             }
           >
@@ -281,6 +287,7 @@ function YAMLEditor({ pattern, onClose, onSubmit }) {
                 id: pattern.id,
                 name: pattern.name,
                 type: FILE_OPS.DELETE,
+                catalog_data: pattern.catalog_data,
               })
             }
           >
@@ -321,6 +328,11 @@ function MesheryPatterns({
 
   const [canPublishPattern, setCanPublishPattern] = useState(false);
   const [publishSchema, setPublishSchema] = useState({});
+  const [infoModal, setInfoModal] = useState({
+    open: false,
+    ownerID: '',
+    selectedResource: {},
+  });
 
   const [viewType, setViewType] = useState('grid');
   const { notify } = useNotification();
@@ -352,6 +364,7 @@ function MesheryPatterns({
   });
 
   const [loading, stillLoading] = useState(true);
+  const { width } = useWindowDimensions();
 
   const catalogVisibilityRef = useRef(false);
   const catalogContentRef = useRef();
@@ -671,6 +684,20 @@ function MesheryPatterns({
     });
   };
 
+  const handleInfoModalClose = () => {
+    setInfoModal({
+      open: false,
+    });
+  };
+
+  const handleInfoModal = (pattern) => {
+    setInfoModal({
+      open: true,
+      ownerID: pattern.user_id,
+      selectedResource: pattern,
+    });
+  };
+
   const handlePublishModal = (ev, pattern) => {
     if (canPublishPattern) {
       ev.stopPropagation();
@@ -877,7 +904,7 @@ function MesheryPatterns({
     };
   }
 
-  async function handleSubmit({ data, id, name, type, metadata }) {
+  async function handleSubmit({ data, id, name, type, metadata, catalog_data }) {
     updateProgress({ showProgress: true });
     if (type === FILE_OPS.DELETE) {
       const response = await showModal(1, name);
@@ -907,7 +934,10 @@ function MesheryPatterns({
         {
           credentials: 'include',
           method: 'POST',
-          body: JSON.stringify({ pattern_data: { id, pattern_file: data }, save: true }),
+          body: JSON.stringify({
+            pattern_data: { id, pattern_file: data, catalog_data },
+            save: true,
+          }),
         },
         () => {
           console.log('PatternFile API', `/api/pattern`);
@@ -925,12 +955,18 @@ function MesheryPatterns({
           pattern_data: {
             name: metadata?.name || name,
             pattern_file: data,
+            catalog_data,
           },
           save: true,
         });
       }
       if (type === FILE_OPS.URL_UPLOAD) {
-        body = JSON.stringify({ url: data, save: true, name: metadata?.name || name });
+        body = JSON.stringify({
+          url: data,
+          save: true,
+          name: metadata?.name || name,
+          catalog_data,
+        });
       }
       dataFetch(
         `/api/pattern`,
@@ -971,6 +1007,14 @@ function MesheryPatterns({
   const handleOpenInConfigurator = (id) => {
     router.push('/configuration/designs/configurator?design_id=' + id);
   };
+
+  let colViews = [
+    ['name', 'xs'],
+    ['created_at', 'm'],
+    ['updated_at', 'l'],
+    ['visibility', 's'],
+    ['Actions', 'xs'],
+  ];
 
   const columns = [
     {
@@ -1084,7 +1128,11 @@ function MesheryPatterns({
           const rowData = patterns[tableMeta.rowIndex];
           const visibility = patterns[tableMeta.rowIndex]?.visibility;
           return (
-            <>
+            <Box
+              sx={{
+                display: 'flex',
+              }}
+            >
               {userCanEdit(rowData) && (
                 <TooltipIcon
                   placement="top"
@@ -1169,6 +1217,10 @@ function MesheryPatterns({
                 <GetAppIcon data-cy="download-button" />
               </TooltipIcon>
 
+              <TooltipIcon title="Design Information" onClick={() => handleInfoModal(rowData)}>
+                <InfoOutlinedIcon data-cy="information-button" />
+              </TooltipIcon>
+
               {canPublishPattern && visibility !== VISIBILITY.PUBLISHED ? (
                 <TooltipIcon
                   placement="bottom"
@@ -1185,7 +1237,7 @@ function MesheryPatterns({
                   <PublicIcon fill="#F91313" data-cy="unpublish-button" />
                 </TooltipIcon>
               )}
-            </>
+            </Box>
           );
         },
       },
@@ -1201,10 +1253,11 @@ function MesheryPatterns({
   const [tableCols, updateCols] = useState(columns);
 
   const [columnVisibility, setColumnVisibility] = useState(() => {
+    let showCols = updateVisibleColumns(colViews, width);
     // Initialize column visibility based on the original columns' visibility
     const initialVisibility = {};
     columns.forEach((col) => {
-      initialVisibility[col.name] = col.options?.display !== false;
+      initialVisibility[col.name] = showCols[col.name];
     });
     return initialVisibility;
   });
@@ -1523,6 +1576,7 @@ function MesheryPatterns({
             setPublishModal={setPublishModal}
             publishSchema={publishSchema}
             user={user}
+            handleInfoModal={handleInfoModal}
           />
         )}
         <ConfirmationModal
@@ -1574,6 +1628,17 @@ function MesheryPatterns({
               />
             }
             submitBtnIcon={<PublishIcon className={classes.addIcon} data-cy="import-button" />}
+          />
+        )}
+        {infoModal.open && (
+          <InfoModal
+            infoModalOpen={true}
+            handleInfoModalClose={handleInfoModalClose}
+            dataName="patterns"
+            selectedResource={infoModal.selectedResource}
+            resourceOwnerID={infoModal.ownerID}
+            currentUserID={user?.id}
+            formSchema={publishSchema}
           />
         )}
         <PromptComponent ref={modalRef} />
